✨ feat(energy): Use Long for energy values
This commit updates the energy system to use `Long` instead of `Integer` for energy values. This change accommodates larger energy capacities and transfers.

Key changes include:

- **Type updates:** `Integer` is replaced with `Long` in various energy-related components, interfaces, and handlers, including `IEnergyTarget`, `IEnergyNetwork`, `EnergyNetwork`, `EnergyTargetCapabilityProvider`, `IngredientComponentValueHandlerEnergy`, `IngredientComponentCapabilities`, `EnergyHelpers`, `TileEnergyBattery`, `TileCoalGenerator`, `TunnelEnergyHelpers`, `IngredientComponentHandlers`, `CraftingAspects`, `CraftingAspectWriteBuilders`, and `IngredientComponentTerminalStorageHandlerEnergy`.
- **Value adjustments:** Methods and fields that previously handled `int` for energy now handle `long`.
- **Capacity handling:** The `receiveEnergy` method in `TileEnergyBattery` now considers `getEnergyPerTick()` to prevent overflow.
- **Shift-click crafting:** A new configuration option `shiftClickCraftingResultLimit` is introduced to control single-item crafting from the result slot.
- **Portable terminal opening:** A new keybinding and packet (`TerminalStorageIngredientItemOpenGenericPacket`) are added for opening portable storage terminals.
- **JEI/NEI integration:** `TerminalStorageRecipeTransferHandler` is updated to be generic for different container types.
- **GUI initialization:** `GuiTerminalStorage` now resets its `initialized` flag correctly.
- **Crafting grid

+++ src/main/java/ruiseki/integrateddynamics/capability/ingredient/IngredientComponentCapabilities.java
@@ -58,13 +58,13 @@ public class IngredientComponentCapabilities {
                 }
             });
         attacherManager.addAttacher(
-            new IngredientComponentCapabilityAttacherAdapter<Integer, Boolean>(
+            new IngredientComponentCapabilityAttacherAdapter<Long, Boolean>(
                 INGREDIENT_ENERGY_NAME,
                 capabilityIngredientComponentValueHandler) {
 
                 @Override
                 public ICapabilityProvider createCapabilityProvider(
-                    IngredientComponent<Integer, Boolean> ingredientComponent) {
+                    IngredientComponent<Long, Boolean> ingredientComponent) {
                     return new DefaultCapabilityProvider<>(
                         () -> IngredientComponentValueHandlerConfig.CAPABILITY,
                         new IngredientComponentValueHandlerEnergy(ingredientComponent));
