+++ src/main/java/ruiseki/integrateddynamics/core/network/EnergyNetwork.java
@@ -11,16 +11,16 @@ import ruiseki.integrateddynamics.api.network.INetworkElement;
 
 /**
  * A network that can hold energy.
- * 
+ *
  * @author rubensworks
  */
-public class EnergyNetwork extends PositionedAddonsNetworkIngredients<Integer, Boolean> implements IEnergyNetwork {
+public class EnergyNetwork extends PositionedAddonsNetworkIngredients<Long, Boolean> implements IEnergyNetwork {
 
     @Getter
     @Setter
     private INetwork network;
 
-    public EnergyNetwork(IngredientComponent<Integer, Boolean> component) {
+    public EnergyNetwork(IngredientComponent<Long, Boolean> component) {
         super(component);
     }
